requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tqdm
numpy
pillow
scikit-learn
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
tensorboard
